Allow toggling whether to purge unsubscribed or not

+++ sources/Odysee.ts
@@ -274,9 +274,12 @@ export default class Odysee extends Source {
 
 	override async postRunTasks(
 		prisma: PrismaClient,
-		subscriptions: { channel: string; types: VideoTypeSelector }[]
+		subscriptions: { channel: string; types: VideoTypeSelector }[],
+		doPurgeUnsubscribed: boolean
 	): Promise<void> {
-		await purgeUnsubscribed(prisma, subscriptions);
+		if (doPurgeUnsubscribed) {
+			await purgeUnsubscribed(prisma, subscriptions);
+		}
 	}
 }
