+++ sources/PeerTube.ts
@@ -254,9 +254,12 @@ export default class PeerTube extends Source {
 
 	override async postRunTasks(
 		prisma: PrismaClient,
-		subscriptions: { channel: string; types: VideoTypeSelector }[]
+		subscriptions: { channel: string; types: VideoTypeSelector }[],
+		doPurgeUnsubscribed: boolean
 	): Promise<void> {
-		await purgeUnsubscribed(prisma, subscriptions);
+		if (doPurgeUnsubscribed) {
+			await purgeUnsubscribed(prisma, subscriptions);
+		}
 	}
 }
